Build changes for OSX
OSX packages need to be notarized but GoReleaser doesn't support post hooks
on the archive action. To get around this we create `build/buildosx.sh` that
runs as a post hook on the build step. It signs the binary, builds a zipfile
of it, and then notarizes the zip. I don't know (and goReleaser provides no
discernable way to test) if the publish step will grab zips that GoReleaser
didn't create. So we'll _do it live!_